pyproject.toml
# SPDX-FileCopyrightText: 2020-2023 CERN
# SPDX-FileCopyrightText: 2023 GSI Helmholtzzentrum für Schwerionenforschung
# SPDX-FileNotice: All rights not expressly granted are reserved.
#
# SPDX-License-Identifier: GPL-3.0-or-later OR EUPL-1.2+

[build-system]
requires = [
    'setuptools >= 61',
    'wheel >= 0.36',
    'setuptools-scm[toml] >= 7.0',
]
build-backend = 'setuptools.build_meta'

[project]
dynamic = ['version']
name = 'acc-app-optimisation'
description = 'GUI for generic numerical optimisation'
readme = 'README.md'
authors = [
    {name = 'Verena Kain', email = 'verena.kain@cern.ch'},
    {name = 'Nico Madysa', email = 'nico.madysa@cern.ch'},
]
requires-python = '>=3.7'
classifiers = [
    'Development Status :: 3 - Alpha',
    'Environment :: X11 Applications :: Qt',
    'Intended Audience :: Science/Research',
    'License :: Other/Proprietary License',
    'Natural Language :: English',
    'Operating System :: POSIX :: Linux',
    'Programming Language :: Python :: 3 :: Only',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Topic :: Scientific/Engineering :: Artificial Intelligence',
    'Topic :: Scientific/Engineering :: Physics',
]
dependencies = [
    'JPype1 ~= 1.0',
    'Py-BOBYQA ~= 1.2',
    'PyQt5 ~= 5.12',
    'accwidgets[app_frame,graph,log_console,lsa_selector,rbac,screenshot] ~= 1.7.5',
    'cern-awake-env ~= 0.18.0',
    'linac3-lebt-tuning ~= 1.0',
    'cern-isolde-offline-env ~= 0.0.5',
    'cern-leir-transfer-line-env ~= 0.5.2',
    'cern-sps-tune-env ~= 0.4.0',
    'cern-sps-splitter-opt-env ~= 0.0.8',
    'cern-sps-zs-alignment-env ~= 0.4.0',
    'cernml-coi ~= 0.8.9',
    'cernml-coi-utils ~= 0.2.6',
    'cernml-svd ~= 3.0',
    'cernml-extremum-seeking ~= 2.0',
    'gym >= 0.17',
    'importlib-metadata >=3.6; python_version<"3.8"',
    'numpy ~= 1.20',
    'pjlsa ~= 0.2.0',
    'pyjapc ~= 2.0',
    'pylogbook ~= 3.3',
    'scikit-optimize ~= 0.9.0',
    'scipy ~= 1.3',
    'stable-baselines3 ~= 1.0',
]

[project.optional-dependencies]
test = [
    'pytest >= 6.0',
    'pytest-mock >= 3.3',
    'pytest-qt >= 4.2',
]
dev = [
    'PyQt5-stubs ~= 5.12',
    'black >= 23',
    'mypy ~= 1.0',
    'pylint ~= 2.0',
    'reuse',
]

[project.urls]
gitlab = 'https://gitlab.cern.ch/geoff/geoff-app'

[tool.setuptools]
zip-safe = true
platforms = [
    'CERN General Network',
    'CERN Technical Network',
]

[tool.setuptools.packages.find]
include = ['acc_app_optimisation*']

[tool.setuptools_scm]

[tool.black]
target-version = ['py37', 'py38', 'py39']

[tool.isort]
profile = 'black'

[tool.pylint.main]
extension-pkg-allow-list = ['PyQt5']

[tool.pylint.reports]
score = false
reports = false
output-format = 'parseable'

[tool.pylint.format]
ignore-long-lines = '<?https?://\S+>?$'

[tool.pylint.design]
max-attributes = 14

[tool.pylint.'messages control']
disable = [
    'duplicate-code',
    'fixme',
    'missing-module-docstring',
    'missing-class-docstring',
    'missing-function-docstring',
    'too-few-public-methods',
]

[tool.mypy]
disallow_untyped_defs = true
namespace_packages = true

[[tool.mypy.overrides]]
module = [
    'accwidgets.*',
    'gym.*',
    'jpype.*',
    'matplotlib.*',
    'numpy.*',
    'scipy.*',
    'skopt.*',
    'pjlsa.*',
    'pybobyqa.*',
    'pyjapc.*',
    'pyqtgraph.*',
    'pytestqt.*',
]
ignore_missing_imports = true
